Cost addet to each turret to purchase - For limiting on placing the turret

+++ Tower Defence_Brackeys_Tutorial/Assets/Scripts/TurretBlueprint.cs
@@ -6,5 +6,5 @@ using UnityEngine;
 public class TurretBlueprint 
 {
     public GameObject turret;
-    public float cost;
+    public int cost;
 }
